+++ prettier.config.js
@@ -1,3 +1,7 @@
 module.exports = {
-    tabWidth: 4
+    tabWidth: 4,
+    singleQuote: false,
+    prettyWidth: 80,
+    trailingComma: "none",
+    useTabs: false
 };