+++ CHANGES.md
@@ -48,6 +48,7 @@
 *   Added tooltip to future dates on datepicker
 *   Header adjustment for new design system.
 *   Updated notification with govau design systems components & fixed a few minor error handling related issues
+*   Implemented SMTP client with `nodemailer` for correspondence api
 *   Simplified the way that search queries are sent to elasticsearch and boosted title to higher importance.
 *   Modified `Show full description` font-size.
 *   Added `recrawl` API endpoints to sleuther framework